requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
langchain==0.1.9
langchain-community==0.0.24
faiss-cpu==1.7.4
sentence-transformers==2.5.1
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
paddleocr==2.7.0.3
groq==0.4.2
sqlalchemy==2.0.27
pydantic==2.6.1 